Redesign empty workspace and improve "opening" experience
resolve #134

"Open" button can open any sort or resource: DME or DMM.
For DME there will be the same flow, as it was.
For DMM we will check if an environment is opened and, if it's not,
we will recursively check dirs from down to up to find a proper DME file.

Recent maps menu item now shows ALL recent maps ever opened. From different environments.

Backported "Available Maps" functionality.