fix: deletion processing out of bounds tiles (#191)
fixes #164

Out of bounds is only allowed for the regular deletion tool which does not use the bound to determine tiles, 
however it is a problem with the alt function and causes an infinite recursion in findChunkBounds. 
This makes the out of bounds allowed condition determined by if the alt function is enabled, 
fixing the issue. Deletion continues to work as expected.