fix: infinite scrolling when the canvas is exited during a scroll (#192)
fixes #147

processCameraZoom() normally exits when there is no mouseScroll value (by checking zoomed, a var set by processMouseScroll()).

However, processMouseScroll() does not execute if the control is not Active(). 
This means if there is a mouseWheel value on the control and it is exited (no longer active) 
e.g. by entering the right-click menu, zoomed will not be set to false, 
meaning processCameraZoom() will continue to execute.

A simple fix for this is to check if the current control is active in processCameraZoom().